add git submodule history rewrite script and orig
A handy - sort of - script to merge a submodule into the source tree in
which it is embedded.

The script also rewrites the history of the submodule and merges in to the history
of the main source tree.

The orginal from https://github.com/jeremysears/scripts/blob/master/bin/git-submodule-rewrite
was modified a bit to work with prrte.  Problem with the original script is there were many
tags and trying to merge into the main source tree history failed.  So that section is
commented out in the updated script.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>